requirements.txt
# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# Google Generative AI
google-generativeai

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.60.1

# Data processing and visualization
yfinance>=0.2.36
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.1

# Tabulate for pretty-printing tables
tabulate

# Utilities
aiohttp==3.9.3
requests>=2.28.0
uuid

# Natural Language Toolkit
nltk>=3.8.1